rename versioned mapper files
