+++ lib/index.js
@@ -7,8 +7,8 @@ var Identify = require('segmentio-facade').Identify;
 var integration = require('segmentio-integration');
 var isostring = require('isostring');
 var hash = require('string-hash');
-var mapperV2 = require('./v2.mapper');
-var mapperV1 = require('./v1.mapper');
+var mapperV2 = require('./mapper-v2');
+var mapperV1 = require('./mapper-v1');
 var time = require('unix-time');
 var extend = require('extend');
 var dot = require('obj-case');
