+++ test/index.js
@@ -3,8 +3,8 @@ var Test = require('segmentio-integration-tester');
 var helpers = require('./helpers');
 var facade = require('segmentio-facade');
 var hash = require('string-hash');
-var mapperV2 = require('../lib/v2.mapper');
-var mapperV1 = require('../lib/v1.mapper');
+var mapperV2 = require('../lib/mapper-v2');
+var mapperV1 = require('../lib/mapper-v1');
 var time = require('unix-time');
 var should = require('should');
 var assert = require('assert');
